# 05_microservicio_alumnos/src/main/resources/openapi/alumnos.yaml
openapi: 3.0.1
info:
  title: API de Alumnos
  version: 1.0.0
  description: API REST para la gestión de alumnos
paths:
  /alumnos:
    post:
      summary: Alta de alumno
      tags: [Alumnos]
      operationId: createAlumno
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Alumno'
      responses:
        '201':
          description: Alumno creado correctamente
        '400':
          description: Datos inválidos

    get:
      summary: Listado de todos los alumnos
      tags: [Alumnos]
      operationId: getAllAlumnos
      responses:
        '200':
          description: Lista de alumnos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alumno'
    put:
      summary: Actualiza nota de curso
      tags: [Alumnos]
      operationId: updateCalificaciones
      parameters:
        - name: curso
          in: query
          required: true
          schema:
            type: string
        - name: valor
          in: query
          required: true
          schema:
            type: double
      responses:
        '200':
          description: Actualizadas calificaciones del curso
        '409':
          description: No se pudo completar la operación
  /cursos:
    get:
      summary: Listado de cursos
      operationId: getCursos
      tags: [Alumnos]
      responses:
        '200':
          description: Lista de cursos
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /alumnos/curso:
    get:
      summary: Listado de alumnos por curso
      operationId: getAlumnosByCurso
      tags: [Alumnos]
      parameters:
        - name: curso
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de alumnos del curso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alumno'

components:
  schemas:
    Alumno:
      type: object
      required:
        - dni
        - nombre
        - calificacion
        - curso
      properties:
        dni:
          type: string
          description: DNI del alumno
          minLength: 1   # evita DNI vacío
          example: "12345678A"

        nombre:
          type: string
          example: "Juan Pérez"

        curso:
          type: string
          example: "Spring Boot"

        calificacion:
          type: number
          format: double
          minimum: 0     # evita notas negativas
          example: 7.5